extend `makeBrowser` in test helpers
