+++ src/background/services/__tests__/helpers.ts
@@ -1,4 +1,5 @@
 import { vi } from 'vitest';
+import type { Browser, Tabs } from 'webextension-polyfill';
 import type { Storage, WalletInfo } from '@/shared/types';
 
 export const walletInfo: WalletInfo = {
@@ -22,6 +23,19 @@ export const makeLogger = () => ({
   log: vi.fn(),
 });
 
+// wraps a real Browser method's own param types with a relaxed (partial)
+// resolved value, so tests can pass minimal tab/window fixtures
+type MockedAsync<Method extends (...args: never[]) => Promise<unknown>> = (
+  ...args: Parameters<Method>
+) => Promise<Partial<Awaited<ReturnType<Method>>>>;
+
+// same, but also allows resolving `undefined` for a "not found" lookup
+type MockedAsyncOptional<
+  Method extends (...args: never[]) => Promise<unknown>,
+> = (
+  ...args: Parameters<Method>
+) => Promise<Partial<Awaited<ReturnType<Method>>> | undefined>;
+
 export const makeBrowser = () => ({
   alarms: {
     onAlarm: { addListener: vi.fn() },
@@ -30,9 +44,49 @@ export const makeBrowser = () => ({
     clear: vi.fn(),
   },
   runtime: {
-    onConnect: { addListener: vi.fn(), removeListener: vi.fn() },
+    getURL: vi.fn<Browser['runtime']['getURL']>(
+      (path) => `chrome-extension://ext-id/${path}`,
+    ),
+    onConnect: {
+      addListener: vi.fn<Browser['runtime']['onConnect']['addListener']>(),
+      removeListener:
+        vi.fn<Browser['runtime']['onConnect']['removeListener']>(),
+    },
+  },
+  windows: {
+    getLastFocused: vi
+      .fn<MockedAsync<Browser['windows']['getLastFocused']>>()
+      .mockResolvedValue({ id: 1 }),
+  },
+
+  tabs: {
+    query: vi
+      .fn<
+        (
+          ...args: Parameters<Browser['tabs']['query']>
+        ) => Promise<Partial<Tabs.Tab>[]>
+      >()
+      .mockResolvedValue([]),
+    get: vi
+      .fn<MockedAsyncOptional<Browser['tabs']['get']>>()
+      .mockResolvedValue(undefined),
+    update: vi
+      .fn<MockedAsync<Browser['tabs']['update']>>()
+      .mockResolvedValue({}),
+    create: vi
+      .fn<MockedAsync<Browser['tabs']['create']>>()
+      .mockResolvedValue({ id: 1 }),
+    remove: vi.fn<Browser['tabs']['remove']>().mockResolvedValue(undefined),
+    highlight: vi
+      .fn<MockedAsync<Browser['tabs']['highlight']>>()
+      .mockResolvedValue({}),
   },
 });
+
+// the fake above only implements the slice of Browser each test needs
+export type FakeBrowser = ReturnType<typeof makeBrowser>;
+export const asBrowser = (browser: FakeBrowser) =>
+  browser as unknown as Browser;
 
 export const makeStorage = (initial: Partial<Storage> = {}) => {
   const state: Partial<Storage> = { ...initial };
